# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TempShield API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  message: string;
}

export interface RegisterRequest {
  name: string;
  email: string;
  password: string;
}

export interface LoginRequest {
  email: string;
  password: string;
}

export type UserProfileRole =
  (typeof UserProfileRole)[keyof typeof UserProfileRole];

export const UserProfileRole = {
  USER: "USER",
  ADMIN: "ADMIN",
} as const;

export type UserProfilePlan =
  (typeof UserProfilePlan)[keyof typeof UserProfilePlan];

export const UserProfilePlan = {
  FREE: "FREE",
  BASIC: "BASIC",
  PRO: "PRO",
} as const;

export interface UserProfile {
  id: number;
  name: string;
  email: string;
  apiKey: string;
  role: UserProfileRole;
  plan: UserProfilePlan;
  requestCount: number;
  requestLimit: number;
  createdAt: string;
}

export interface AuthResponse {
  user: UserProfile;
  message: string;
}

export interface CheckEmailRequest {
  email: string;
}

export interface CheckEmailResponse {
  isDisposable: boolean;
  domain: string;
  reputationScore: number;
  requestsRemaining: number;
  mxValid?: boolean;
  inboxSupport?: boolean;
}

export interface UsageEntry {
  id: number;
  endpoint: string;
  email?: string | null;
  domain?: string | null;
  isDisposable?: boolean | null;
  reputationScore?: number | null;
  timestamp: string;
}

export interface UsageByDay {
  date: string;
  count: number;
}

export interface DashboardData {
  user: UserProfile;
  recentUsage: UsageEntry[];
  usageByDay: UsageByDay[];
}

export interface UsageHistory {
  entries: UsageEntry[];
  total: number;
}

export interface ApiKeyResponse {
  apiKey: string;
  message: string;
}

export type UpgradeRequestPlan =
  (typeof UpgradeRequestPlan)[keyof typeof UpgradeRequestPlan];

export const UpgradeRequestPlan = {
  BASIC: "BASIC",
  PRO: "PRO",
} as const;

export interface UpgradeRequest {
  plan: UpgradeRequestPlan;
  note?: string;
}

export type AdminUserRole = (typeof AdminUserRole)[keyof typeof AdminUserRole];

export const AdminUserRole = {
  USER: "USER",
  ADMIN: "ADMIN",
} as const;

export type AdminUserPlan = (typeof AdminUserPlan)[keyof typeof AdminUserPlan];

export const AdminUserPlan = {
  FREE: "FREE",
  BASIC: "BASIC",
  PRO: "PRO",
} as const;

export interface AdminUser {
  id: number;
  name: string;
  email: string;
  role: AdminUserRole;
  plan: AdminUserPlan;
  requestCount: number;
  requestLimit: number;
  createdAt: string;
}

export interface AdminUsersResponse {
  users: AdminUser[];
  total: number;
}

export type UpdatePlanRequestPlan =
  (typeof UpdatePlanRequestPlan)[keyof typeof UpdatePlanRequestPlan];

export const UpdatePlanRequestPlan = {
  FREE: "FREE",
  BASIC: "BASIC",
  PRO: "PRO",
} as const;

export interface UpdatePlanRequest {
  plan: UpdatePlanRequestPlan;
  note?: string;
}

export type UpgradeRequestItemPlanRequested =
  (typeof UpgradeRequestItemPlanRequested)[keyof typeof UpgradeRequestItemPlanRequested];

export const UpgradeRequestItemPlanRequested = {
  BASIC: "BASIC",
  PRO: "PRO",
} as const;

export type UpgradeRequestItemStatus =
  (typeof UpgradeRequestItemStatus)[keyof typeof UpgradeRequestItemStatus];

export const UpgradeRequestItemStatus = {
  PENDING: "PENDING",
  APPROVED: "APPROVED",
  REJECTED: "REJECTED",
} as const;

export interface UpgradeRequestItem {
  id: number;
  userId: number;
  userName: string;
  userEmail: string;
  planRequested: UpgradeRequestItemPlanRequested;
  status: UpgradeRequestItemStatus;
  note?: string;
  createdAt: string;
}

export interface UpgradeRequestsResponse {
  requests: UpgradeRequestItem[];
  total: number;
}

export type UpdateUpgradeRequestBodyStatus =
  (typeof UpdateUpgradeRequestBodyStatus)[keyof typeof UpdateUpgradeRequestBodyStatus];

export const UpdateUpgradeRequestBodyStatus = {
  APPROVED: "APPROVED",
  REJECTED: "REJECTED",
} as const;

export interface UpdateUpgradeRequestBody {
  status: UpdateUpgradeRequestBodyStatus;
  note?: string;
}

export interface SyncResponse {
  message: string;
  domainsAdded: number;
  totalDomains: number;
}

export type AdminStatsUsersByPlan = {
  FREE: number;
  BASIC: number;
  PRO: number;
};

export interface AdminStats {
  totalUsers: number;
  totalApiCalls: number;
  totalDomains: number;
  pendingUpgradeRequests: number;
  usersByPlan: AdminStatsUsersByPlan;
}